requirements.txt
fastapi
uvicorn
openai
chromadb
sentence-transformers
pymupdf
python-dotenv
